feat(diagram-maker): multi-node alignment + gate controls to single selection
- hide node/edge property controls (type, icon, color scheme, handles,
  etc.) when more than one node or edge is selected, since these are
  single-target edits
- add alignment controls (left/center/right, top/middle/bottom) shown
  when 2+ nodes are selected; computes bounding box from node position
  + measured/data size and repositions selected nodes, edges untouched
- extract DiagramEditorContext.tsx into useNodeSelectionControls,
  useEdgeSelectionControls, and useLocalDiagrams hooks to keep the
  provider file manageable
