+++ apps/diagram-maker/src/app/components/Sidebar/Sidebar.tsx
@@ -13,11 +13,13 @@ import { EdgeActiveControl } from './EdgeProps/EdgeActiveControl';
 import { NodeTypeControl } from './NodeProps/NodeTypeControl';
 import { IconControl } from './NodeProps/IconControl';
 import { TransparentBackgroundControl } from './NodeProps/TransparentBackgroundControl';
+import { AlignmentControls } from './MultiNodeProps/AlignmentControls';
 
 export function Sidebar() {
   const { selectedNodeIds, selectedEdgeIds } = useDiagramEditor();
-  const hasNodeSelection = selectedNodeIds.length > 0;
-  const hasEdgeSelection = selectedEdgeIds.length > 0;
+  const hasSingleNodeSelection = selectedNodeIds.length === 1;
+  const hasMultiNodeSelection = selectedNodeIds.length > 1;
+  const hasSingleEdgeSelection = selectedEdgeIds.length === 1;
 
   return (
     <aside className="w-64 h-full bg-neutral-800 border-r border-neutral-300 p-4 flex flex-col gap-4 overflow-y-auto">
@@ -26,10 +28,10 @@ export function Sidebar() {
 
       <SelectionInfo />
 
-      {/* Node-specific controls */}
+      {/* Node-specific controls (single-node selection only) */}
       <div
         className={`transition-all duration-300 ease-in-out flex flex-col gap-4 ${
-          hasNodeSelection
+          hasSingleNodeSelection
             ? 'opacity-100'
             : 'opacity-0 max-h-0 overflow-hidden pointer-events-none'
         }`}
@@ -42,10 +44,21 @@ export function Sidebar() {
         <LayerControls />
       </div>
 
-      {/* Edge-specific controls */}
+      {/* Multi-node controls (2+ nodes selected, regardless of edge selection) */}
       <div
         className={`transition-all duration-300 ease-in-out flex flex-col gap-4 ${
-          hasEdgeSelection
+          hasMultiNodeSelection
+            ? 'opacity-100'
+            : 'opacity-0 max-h-0 overflow-hidden pointer-events-none'
+        }`}
+      >
+        <AlignmentControls />
+      </div>
+
+      {/* Edge-specific controls (single-edge selection only) */}
+      <div
+        className={`transition-all duration-300 ease-in-out flex flex-col gap-4 ${
+          hasSingleEdgeSelection
             ? 'opacity-100'
             : 'opacity-0 max-h-0 overflow-hidden pointer-events-none'
         }`}
